fix 32-bit failing tests

+++ src/fifobuffer.jl
@@ -89,7 +89,7 @@ Base.wait(f::FIFOBuffer) = wait(f.cond)
 Base.read(f::FIFOBuffer) = readavailable(f)
 Base.flush(f::FIFOBuffer) = nothing
 Base.position(f::FIFOBuffer) = f.f, f.l, f.nb
-function Base.seek(f::FIFOBuffer, pos::Tuple{Int, Int, Int})
+function Base.seek(f::FIFOBuffer, pos::Tuple{Int64, Int64, Int64})
     f.f = pos[1]
     f.l = pos[2]
     f.nb = pos[3]
